+++ front_end/src/store/modules/category/getters.js
@@ -1,0 +1,6 @@
+const getters = {
+  available_categories: (state) => state.available_categories,
+  selected_categories: (state) => state.selected_categories,
+}
+
+export default getters;
