+++ front_end/src/store/modules/category/state.js
@@ -1,0 +1,6 @@
+const state = {
+  available_categories: [],
+  selected_categories:[]
+}
+
+export default state;
