+++ front_end/src/store/modules/order/actions.js
@@ -1,0 +1,11 @@
+import axios from 'axios';
+const actions = {
+  getOrders({ commit }) {
+    axios.get('/api/orders').then((response) => {
+      commit('UPDATE_ORDERS', response.data);
+    });
+  },
+};
+
+
+export default actions;
