+++ front_end/src/store/modules/order/getters.js
@@ -1,0 +1,7 @@
+
+const getters = {
+  orders: (state) => state.orders,
+};
+
+
+export default getters;
