+++ front_end/src/store/modules/order/mutations.js
@@ -1,0 +1,8 @@
+const mutations = {
+  UPDATE_ORDERS(state, payload) {
+    state.orders = payload;
+  },
+};
+
+
+export default mutations;
