+++ front_end/src/store/modules/order/state.js
@@ -1,0 +1,5 @@
+const state = {
+  orders: [],
+};
+
+export default state;
