+++ front_end/src/store/modules/product/actions.js
@@ -1,0 +1,11 @@
+import axios from 'axios';
+
+const actions = {
+  getProductItems({ commit }) {
+    axios.get('/api/products').then((response) => {
+      commit('UPDATE_PRODUCT_ITEMS', response.data);
+    });
+  },
+};
+
+export default actions;
