+++ front_end/src/store/modules/product/mutations.js
@@ -1,0 +1,7 @@
+const mutations = {
+  UPDATE_PRODUCT_ITEMS(state, payload) {
+    state.productItems = payload;
+  },
+};
+
+export default mutations
