+++ front_end/src/store/modules/product/state.js
@@ -1,0 +1,6 @@
+const state = {
+  productItems: [],
+};
+
+
+export default state;
